Rename all ast files to nodes
